+++ R/read_hpa_spatial_data.R
@@ -77,7 +77,8 @@ get_data_file_list <- function() {
     "lhn",
     "phn",
     "mb21_poly",
-    "mb21_pop"
+    "mb21_pop",
+    "mmm19"
   )
 }